# packages/engine/paima-rest/src/tsoa/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"FailedResult": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					},
					"errorMessage": {
						"type": "string"
					},
					"errorCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"errorMessage"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"InternalServerErrorResult": {
				"$ref": "#/components/schemas/FailedResult"
			},
			"FieldErrors": {
				"description": "comes from the `tsoa` package, but we don't want it as a dependency just for this type",
				"properties": {},
				"type": "object",
				"additionalProperties": {
					"properties": {
						"value": {},
						"message": {
							"type": "string"
						}
					},
					"required": [
						"message"
					],
					"type": "object"
				}
			},
			"ValidateErrorResult": {
				"properties": {
					"message": {
						"type": "string",
						"enum": [
							"Validation Failed"
						],
						"nullable": false
					},
					"details": {
						"$ref": "#/components/schemas/FieldErrors"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Achievement": {
				"properties": {
					"name": {
						"type": "string",
						"description": "Unique Achievement String"
					},
					"score": {
						"type": "number",
						"format": "double",
						"description": "Optional: Relative Value of the Achievement"
					},
					"category": {
						"type": "string",
						"description": "Optional: 'Gold' | 'Diamond' | 'Beginner' | 'Advanced' | 'Vendor'"
					},
					"percentCompleted": {
						"type": "number",
						"format": "double",
						"description": "Percent of players that have unlocked the achievement"
					},
					"isActive": {
						"type": "boolean",
						"description": "If achievement can be unlocked at the time."
					},
					"displayName": {
						"type": "string",
						"description": "Achievement Display Name"
					},
					"description": {
						"type": "string",
						"description": "Achievement Description"
					},
					"spoiler": {
						"type": "string",
						"enum": [
							"all",
							"description"
						],
						"description": "Hide entire achievement or description if not completed"
					},
					"iconURI": {
						"type": "string",
						"description": "Optional Icon for Achievement"
					},
					"iconGreyURI": {
						"type": "string",
						"description": "Optional Icon for locked Achievement"
					},
					"startDate": {
						"type": "string",
						"description": "Optional Date ISO8601 YYYY-MM-DDTHH:mm:ss.sssZ"
					},
					"endDate": {
						"type": "string",
						"description": "Optional Date ISO8601 YYYY-MM-DDTHH:mm:ss.sssZ"
					}
				},
				"required": [
					"name",
					"isActive",
					"displayName",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AchievementPublicList": {
				"description": "Result of \"Get All Available Achievements\"",
				"properties": {
					"id": {
						"type": "string",
						"description": "Game ID"
					},
					"name": {
						"type": "string",
						"description": "Optional game name"
					},
					"version": {
						"type": "string",
						"description": "Optional game version"
					},
					"block": {
						"type": "number",
						"format": "double",
						"description": "Data block height (0 always valid)"
					},
					"caip2": {
						"type": "string",
						"description": "CAIP-2 blockchain identifier"
					},
					"time": {
						"type": "string",
						"description": "Optional date. ISO8601, like YYYY-MM-DDTHH:mm:ss.sssZ"
					},
					"achievements": {
						"items": {
							"$ref": "#/components/schemas/Achievement"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"block",
					"caip2",
					"achievements"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PlayerAchievement": {
				"properties": {
					"name": {
						"type": "string",
						"description": "Unique Achievement String"
					},
					"completed": {
						"type": "boolean",
						"description": "Is Achievement completed"
					},
					"completedDate": {
						"type": "string",
						"format": "date-time",
						"description": "Completed Date ISO8601 YYYY-MM-DDTHH:mm:ss.sssZ"
					},
					"completedRate": {
						"properties": {
							"total": {
								"type": "number",
								"format": "double",
								"description": "Total Progress"
							},
							"progress": {
								"type": "number",
								"format": "double",
								"description": "Current Progress"
							}
						},
						"required": [
							"total",
							"progress"
						],
						"type": "object",
						"description": "If achievement has incremental progress"
					}
				},
				"required": [
					"name",
					"completed"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PlayerAchievements": {
				"description": "Result of \"Get Completed Achievements\"",
				"properties": {
					"block": {
						"type": "number",
						"format": "double",
						"description": "Data block height (0 always valid)"
					},
					"caip2": {
						"type": "string",
						"description": "CAIP-2 blockchain identifier"
					},
					"time": {
						"type": "string",
						"description": "Optional date. ISO8601, like YYYY-MM-DDTHH:mm:ss.sssZ"
					},
					"wallet": {
						"type": "string",
						"description": "e.g. addr1234... or 0x1234..."
					},
					"walletType": {
						"type": "string",
						"description": "(Optional) Wallet-type"
					},
					"userId": {
						"type": "string",
						"description": "(Optional) User ID for a specific player account. This value should be\nimmutable and define a specific account, as the wallet might be migrated\nor updated."
					},
					"userName": {
						"type": "string",
						"description": "(Optional) Player display name"
					},
					"completed": {
						"type": "number",
						"format": "double",
						"description": "Total number of completed achievements for the game"
					},
					"achievements": {
						"items": {
							"$ref": "#/components/schemas/PlayerAchievement"
						},
						"type": "array"
					}
				},
				"required": [
					"block",
					"caip2",
					"wallet",
					"completed",
					"achievements"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "@paima/rest",
		"version": "1.0.0",
		"description": "Paima Engine REST, defining the REST endpoints for the Paima node",
		"contact": {
			"name": "Paima Studios"
		}
	},
	"paths": {
...
		"/achievements/public/list": {
			"get": {
				"operationId": "AchievementsPublic_list",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AchievementPublicList"
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorResult"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerErrorResult"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"501": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailedResult"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "category",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "isActive",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "header",
						"name": "Accept-Language",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/achievements/wallet/{wallet}": {
			"get": {
				"operationId": "AchievementsWallet",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PlayerAchievements"
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorResult"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerErrorResult"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "wallet",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Comma-separated list.",
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/achievements/erc/{erc}/{cde}/{token_id}": {
			"get": {
				"operationId": "AchievementsNft",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PlayerAchievements"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailedResult"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorResult"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerErrorResult"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"501": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailedResult"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "erc",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "cde",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "token_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}